+++ packages/react/src/index.ts
@@ -16,7 +16,7 @@ export { default as Attachment, AttachmentProps } from './components/Attachment/
 
 export { default as Avatar, AvatarProps } from './components/Avatar/Avatar'
 
-export { default as Box, BoxProps } from './components/Box/Box'
+export { default as Box, BoxProps } from './components/Box/BoxHeavy'
 
 export { default as Button, ButtonProps, ButtonState } from './components/Button/Button'
 export { default as ButtonGroup, ButtonGroupProps } from './components/Button/ButtonGroup'
